refactor: move around getResolvedEnv and config builder logic
